Fix `SubscribeChannel` type issues
The current version of `SubscribeChannel` type has some issues that make
it difficult for Typescript to narrow types correctly.

Switch to a union type for with-url and without-url cases to fix the issue.
Also remove `SubscribeChannelFunction` type as it's no longer valid in
favor of `SubscribeChannel['func']` type.

Apply the required modifications to the code that used this the type:
`Dialer` and `CallbackUtils`. Introduce a `hasUrl` private method to
`Dialer` in order to narrow some types (needed to apply modifications).