Prevent connecting to discovered nodes multiple times in `Dialer`
When we call `dialProtocol` for a node, it returns a promise. Before the
promise is resolved, if the same node gets discovered again (which is a
possible, because the current versions of guards advertise themselves
every one second), `dialProtocol` is called again, resulting in multiple
connections to the same node, which other than the first one, all are
redundant.

Introduce a pending dial mechanism in which when we start dialing a node,
we save the peer of node and in subsequent discovers first check if the
same peer is trying to dial and if so, prevent dialing again.