Implement a message queue in `Dialer`
When we try to send messages in `Dialer`, it may cause a large number of
streams to be opened to a node and causing some unexpected results, as
multiple calls to `sendMessage` may occur while other ones are trying to
establish connections or streams to a remote node.

To prevent this, introduce a message queue, which is nothing more than a
simple stream. Whenever we want to send a message to a peer, we just push
it to the queue. A job is started when an instance is created, which loops
over this queue and if there is any message in it, it pipes it to its
corresponding stream and so sending it where it should be sent.

If any message in the queue cannot be sent successfully, we retry sending
message with an interval which increases exponentially. After some retries,
if it still fails, the message is dropped totally.

In addition:
- Sort imports
- Remove `streamForPeer` method and add `pushMessageToMessageQueue`
- Rename `getOpenStream` to `getOpenStreamAndConnection`, fix some of its
  issues and switch from checking connection directions to checking stream
  directions