Fix issue in pending dials of `Dialer`
Push new peers to `pendingDialPeers` in `addAndDialPeers` instead of
discovery event handler. If we don't do that, same peers may be added to
the array multiple times, which causes bugs.

In addition, remove pending dials both if the dial is successful or not,
as opposed to current version which doesn't take into account the failed
cases.