Convert `sendMessage` of `Dialer` to a normal function
There is an issue in `ts-mockito` which causes it not to be able to mock
a class field of type async arrow function correctly. It mocks
`sendMessage` to an object, which is clearly wrong and causes the tests
to crash.

Temporarily convert the function to a normal one, preventing such errors
in `ts-mockito`.