Do not dial peers in `addAndDialPeer` of `Dialer` manually
Because `autoDial` option of `libp2p` is enabled, there is no need to
dial the peers manually: they will be dialed automatically.

In addition:
- Rename `addAndDialPeer` to `addPeerToAddressBook` and update js docs
- Switch from `set` method of address book to `add`, as it is preferred
  based on the docs

Note: `libp2p` migration guide for version `0.40.0` states that `autoDial`
option of `libp2p` no longer works as expected and we should dial to all
discovered peers manually, but many issues occur if we try to disable the
option. For example, if we disable it, the discovery mechanism doesn't
work anymore and no guard will be able to find any other guard. Until
these kinds of issues are not resolved, we should stick to `autoDial`.