Delete disconnected peers from address book in `Dialer`
It seems that the `autoDial` option of `libp2p` only works the first time
a peer is added to address book. In other words, the peer is only dialed
the first time it is added to address book. If the peer, for example, is
disconnected (as a result of a hop reset, in our case), it will not be
dialed again all the times.

In order to fix the issue, we need to remove the peer from address book
whenever the peer is disconnected and add it to the address book again
when it reconnects.