Add `minConnections` config in `Dialer`
Guards try to connect to new guards until they reach `minConnection` limit
of `libp2p` config. When the limit is reached, they won't try to connect
to new guards anymore, which is not what we want. If the number of guards
is large, say 30 (which is our case), after connecting to some of them,
they stop connecting to new ones.

Increase the limit to 40 by introducing a new `guardsCount` config with
a value of 30, and adding 10 to it. Please note that if the number of
guards is changed in the future, the config needs to be updated, too.