Add a public key change handling mechanism to `MultiSig`
The public keys which are used to instantiate a `MultiSig` are dynamic
and change over time. The current version of `MultiSig`, however, does
not support the changed values of public keys and use the old ones, which
is wrong.

Add a `handlePublicKeysChange` method to `MultiSig` which is used to
apply required changes whenever the public keys are changed. This method
is called in `GuardConfig.setConfig`, whenever a new set of public keys
are fetched.

In addition:
- Make `MultiSig.getInstance` `publicKeys` parameter optional, so we
  don't have to pass it if we know that it is instantiated before. Also
  add a new overload for the `getInstance`.
- Sort imports in `GuardConfig`
- Update imports order in `fixture` so the mocking for tests work as
  expected
- Add the new format of parameters of `MultiSig.getInstance` to
  `MockedMultiSig`