Utilize child loggers to add file names to logs
In order to increase the readability of logs, we need to add some info
about the file in which the log was called in. Add file name info to all
logs by using `winston` child loggers.

In addition, remove `logger` function export from `Logger` in favor of
`loggerFactory`, which is used to created child loggers with file names.
Also update all affected files to use `loggerFactory` and create their
own loggers.