Implement a relay re-connection mechanism
If a relay is disconnected, there is no way to re-connect to it in the
current source code. In addition, there is no api in js-libp2p to help us
with this issue (e.g. in @libp2p/bootstrap).

To fix the issue, whenever a peer disconnects, check if the peer belongs
to a relay and if so, try to re-connect to it by adding it to the address
book.

In addition:
- Extract checking if a peer belongs to a relay to two separate methods,
  `isPeerTypeRelay` and `isPeerTypeListener`
- Improve peer disconnection logs to clarify if the peer is relay or not