Remove `catch()` expression in `NodeApi`
The `catch` expression causes the returned promise of `sendTx` to resolve
even when we have an error. Because this behavior is unexpected based on
the usage of `sendTx` in `ErgoChain.submitTransaction`, we need to reject
the promise if such an error happens. With current implementation, the
catch block in `ErgoChain.submitTransaction` is never reached in case of
an api error.

Fix the issue by removing the `catch` expression, causing the error handling
work as expected in `ErgoChain.submitTransaction`.