Fix issue in `InputBoxes.getEventValidCommitments`
The current implementation relies on some imperative method for making
the commitments unique. The problem is that, it mutates `event.WIDs` when
its length is 9. This mutation causes issue when creating reward tx, where
the extra commitment assets is considered in candidate output boxes, which
is wrong.

Fix the issue by switching to a declarative, functional, non-mutating
implementation for choosing the correct commitments.