Fix revenue job and revenueView
Previously we used a extractTransactionOrder to extract reward tx orders
and store the related revenues, but it doesn't work for a signed tx. Then
we updated the job to use extractSignedTransactionOrder and work fine.

Revenue view had minor bugs with tables joining, and bug with duplicated
triggers, which have been fixed.