fix: remove pipe medium between tss and guard services
In `child_process.spawn`, there is a size limit for the default medium
between parent and child processes (which is a pipe). If the io size
between the processes is huge and the other side of the pipe doesn't
consume the data, the services may stop working correctly.

Fix the issue by using `ignore` as `stdio` for `spawn`, which in simple
words, redirects child process io to `/dev/null`.